Abstracted Groups and Search
This push refactored the left menu bar by abstracted away the groups and search functionality in it.
